[nrf528xx] Change directory structure for nrf528xx family
