+++ examples/platforms/Makefile.platform.am
@@ -74,11 +74,11 @@ include $(top_srcdir)/examples/platforms/kw41z/Makefile.platform.am
 endif
 
 if OPENTHREAD_EXAMPLES_NRF52811
-include $(top_srcdir)/examples/platforms/nrf52811/Makefile.platform.am
+include $(top_srcdir)/examples/platforms/nrf528xx/nrf52811/Makefile.platform.am
 endif
 
 if OPENTHREAD_EXAMPLES_NRF52840
-include $(top_srcdir)/examples/platforms/nrf52840/Makefile.platform.am
+include $(top_srcdir)/examples/platforms/nrf528xx/nrf52840/Makefile.platform.am
 endif
 
 if OPENTHREAD_EXAMPLES_QPG6095
