+++ examples/platforms/nrf528xx/README.md
