+++ examples/platforms/nrf528xx/nrf52811/openthread-core-nrf52811-config.h
