+++ examples/platforms/nrf528xx/nrf52811/platform-config.h
