+++ examples/platforms/nrf528xx/nrf52840/nrf52840_bootloader.ld
