+++ examples/platforms/nrf528xx/nrf52840/openthread-core-nrf52840-config.h
