+++ examples/platforms/nrf528xx/nrf52840/platform-config.h
