+++ examples/platforms/nrf528xx/src/alarm.c
