+++ examples/platforms/nrf528xx/src/diag.c
