+++ examples/platforms/nrf528xx/src/entropy.c
