+++ examples/platforms/nrf528xx/src/flash.c
