+++ examples/platforms/nrf528xx/src/flash_nosd.c
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2019, The OpenThread Authors.
+ *  Copyright (c) 2017, The OpenThread Authors.
  *  All rights reserved.
  *
  *  Redistribution and use in source and binary forms, with or without
@@ -26,39 +26,33 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  */
 
-/**
- * @file
- *   This file implements the OpenThread FEM helper functions.
- *
- */
-
 #include <openthread-core-config.h>
 #include <openthread/config.h>
 
+#include <assert.h>
 #include <stdint.h>
 #include <string.h>
 
-#include "platform-fem.h"
+#include <utils/code_utils.h>
+#include <utils/flash.h>
 
-#define ENABLE_FEM 1
-#include <nrf_802154.h>
+#include "platform-nrf5.h"
+#include "hal/nrf_nvmc.h"
 
-void PlatformFemSetConfigParams(const PlatformFemConfigParams *aConfig)
+otError nrf5FlashPageErase(uint32_t aAddress)
 {
-    nrf_802154_fem_control_cfg_t cfg;
+    nrf_nvmc_page_erase(aAddress);
 
-    memset(&cfg, 0, sizeof(cfg));
+    return OT_ERROR_NONE;
+}
 
-    cfg.pa_cfg.enable       = aConfig->mPaCfg.mEnable;
-    cfg.pa_cfg.active_high  = aConfig->mPaCfg.mActiveHigh;
-    cfg.pa_cfg.gpio_pin     = aConfig->mPaCfg.mGpioPin;
-    cfg.lna_cfg.enable      = aConfig->mLnaCfg.mEnable;
-    cfg.lna_cfg.active_high = aConfig->mLnaCfg.mActiveHigh;
-    cfg.lna_cfg.gpio_pin    = aConfig->mLnaCfg.mGpioPin;
-    cfg.ppi_ch_id_clr       = aConfig->mPpiChIdClr;
-    cfg.ppi_ch_id_set       = aConfig->mPpiChIdSet;
-    cfg.pa_gpiote_ch_id     = aConfig->mGpiotePaChId;
-    cfg.lna_gpiote_ch_id    = aConfig->mGpioteLnaChId;
+bool nrf5FlashIsBusy(void)
+{
+    return NRF_NVMC->READY != NVMC_READY_READY_Ready;
+}
 
-    nrf_802154_fem_control_cfg_set(&cfg);
+uint32_t nrf5FlashWrite(uint32_t aAddress, const uint8_t *aData, uint32_t aSize)
+{
+    nrf_nvmc_write_bytes(aAddress, aData, aSize);
+    return aSize;
 }
