+++ examples/platforms/nrf528xx/src/logging.c
