+++ examples/platforms/nrf528xx/src/platform-nrf5.h
