+++ examples/platforms/nrf528xx/src/platform-softdevice.h
