+++ examples/platforms/nrf528xx/src/softdevice.h
