+++ examples/platforms/nrf528xx/src/spi-slave.c
