+++ examples/platforms/nrf528xx/src/temp.c
