+++ examples/platforms/nrf528xx/src/uart.c
