+++ examples/platforms/nrf528xx/src/usb-cdc-uart.c
